backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
sqlalchemy>=2.0.0,<3.0.0
qdrant-client==1.11.2
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.0.0
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0
pyyaml==6.0.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0

